Add back button and silhouette to homepage
