+++ src/App.css
@@ -136,8 +136,4 @@ html {
     max-width: 90%;
     margin: 0 auto;
   }
-
-  .c-icon {
-    display: none;
-  }
 }
