+++ src/App.jsx
@@ -5,11 +5,13 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import Continents from './components/Continents';
 import Countries from './components/Countries';
 import Homepage from './components/Homepage';
+import Header from './components/Header';
 
 function App() {
   return (
     <div className="bg-blue-light">
       <HashRouter basename="/">
+        <Header />
         <Routes>
           <Route path="/" element={<Homepage />} />
           <Route path="/continents/:name" element={<Continents />} />
