+++ src/components/Continents.jsx
@@ -1,5 +1,5 @@
 import { useSelector } from 'react-redux';
-import { Link, useParams, useNavigate } from 'react-router-dom';
+import { Link, useParams } from 'react-router-dom';
 import { useState } from 'react';
 import sAmerica from '../img/south-america.svg';
 import nAmerica from '../img/north-america.svg';
@@ -43,18 +43,9 @@ const Continents = () => {
   const continent = objContinents.filter((item) => item.continentName === name);
   const nameC = continent.map((item) => item.continentName);
   const imgC = continent.map((item) => item.silhouette);
-  const navigate = useNavigate();
 
   return (
     <div className="bg-first">
-      <button
-        type="button"
-        className="btn-back c-icon p-1"
-        onClick={() => navigate(-1)}
-      >
-        <box-icon name="left-arrow" animation="burst-hover" color="#ffffff">{ }</box-icon>
-      </button>
-
       <div className="d-flex justify-content-center align-items-center py-5 gap-5 bg-second">
         <img width="200px" src={imgC} alt="" />
         <p className="f-gill fs">{nameC}</p>
