--- src/components/Silhoutte.jsx
@@ -1,39 +1,0 @@
-import sAmerica from '../img/south-america.svg';
-import nAmerica from '../img/north-america.svg';
-import asia from '../img/asia.svg';
-import africa from '../img/africa.svg';
-import europe from '../img/europe.svg';
-import oceania from '../img/australia.svg';
-
-const Silhoutte = () => {
-  const objContinents = [
-    {
-      continentName: 'North America',
-      silhouette: nAmerica,
-    },
-    {
-      continentName: 'South America',
-      silhouette: sAmerica,
-    },
-    {
-      continentName: 'Asia',
-      silhouette: asia,
-    },
-    {
-      continentName: 'Africa',
-      silhouette: africa,
-    },
-    {
-      continentName: 'Europe',
-      silhouette: europe,
-    },
-    {
-      continentName: 'Australia-Oceania',
-      silhouette: oceania,
-    },
-  ];
-
-  return (
-    <div></div>
-  );
-}
